% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuonen.R
\name{kuonen}
\alias{kuonen}
\title{Kuonen's method to calculate p-value}
\usage{
kuonen(q, lambda, tol = 1e-08, maxit = 1000)
}
\arguments{
\item{q}{observed test statistic}

\item{lambda}{eigen-values of the empirical covariance matrix}

\item{tol}{tolorence level}

\item{maxit}{maximum number of iterations}
}
\description{
Kuonen's method to calculate p-value
}
