% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srat.R
\name{obj.min}
\alias{obj.min}
\title{Objective function minimization}
\usage{
obj.min(
  coef,
  Z,
  y,
  u = NULL,
  v = NULL,
  tree = TRUE,
  nrep = 1,
  ftol = 1e-06,
  nfun_max = 5000
)
}
\arguments{
\item{coef}{initial parameter values}

\item{Z}{covariates matrix}

\item{y}{response vector}

\item{u}{a vector of weights}

\item{v}{a vector of weigths}

\item{tree}{indicator of whether the tree algorithm should be used for fast ranking}

\item{nrep}{number of replications in the optimization algorithm}

\item{ftol}{function tolerance to determine convergence}

\item{nfun_max}{maximum number of function evaluations}
}
\description{
Objective function minimization
}
