% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{SRAT-package}
\alias{SRAT-package}
\alias{SRAT}
\title{Sequence Robust Association Test (SRAT)}
\description{
SRAT (Dai, Yang, et al. 2017) <doi:10.1111/biom.12643> is a fully rank-based and 
flexible approach to test for association between a set of genetic variants and an outcome, 
while accounting for within-family correlation and adjusting for covariates. Comparing to 
existing methods, SRAT has the advantages of allowing for unknown correlation structures 
and weaker assumptions about the outcome distribution.
}
\keyword{package}
