% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srat.R
\name{srat.null}
\alias{srat.null}
\title{SRAT with adjustment for additional covariates (null model)}
\usage{
srat.null(y, X, B = 1000)
}
\arguments{
\item{y}{outcome vector (e.g., phenotype)}

\item{X}{covariates for adjustment (e.g., age, gender)}

\item{B}{number of perturbations}
}
\description{
SRAT with adjustment for additional covariates (null model)
}
