% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srat.R
\name{srat}
\alias{srat}
\title{SRAT with adjustment for additional covariates}
\usage{
srat(Y, Z, X, cluster = NULL, init = NULL, w_sqrt = NULL,
  bw = NULL, bw_adj = NULL, n_ptb = 1000)
}
\arguments{
\item{Y}{outcome vector (e.g., phenotype)}

\item{Z}{covariates for testing (e.g., genotype)}

\item{X}{covariates for adjustment (e.g., age, gender)}

\item{cluster}{family identifier}

\item{init}{initial parameters}

\item{w_sqrt}{square root of weight vector}

\item{bw}{bandwidth}

\item{bw_adj}{adjusted factor to bandwidth}

\item{n_ptb}{number of perturbations}
}
\description{
SRAT with adjustment for additional covariates
}
