% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srat0.R
\name{srat0}
\alias{srat0}
\title{SRAT without adjustment for additional covariates}
\usage{
srat0(Y, Z, n_ptb = 1000)
}
\arguments{
\item{Y}{outcome variable}

\item{Z}{covariates for testing}

\item{n_ptb}{number of perturbations}
}
\description{
SRAT without adjustment for additional covariates
}
\examples{
# Data from wilcox.test
x <- c(0.80, 0.83, 1.89, 1.04, 1.45, 1.38, 1.91, 1.64, 0.73, 1.46)
y <- c(1.15, 0.88, 0.90, 0.74, 1.21)
wilcox.test(x, y)
# SRAT p-value
set.seed(123)
Y <- c(x, y)
Z <- rep(0:1, c(length(x), length(y)))
srat0(Y, Z)$p_value

}
