% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srat.R
\name{srat.test}
\alias{srat.test}
\title{SRAT with adjustment for additional covariates (testing)}
\usage{
srat.test(obj, Z, w.sqrt = NULL)
}
\arguments{
\item{obj}{object from null model}

\item{Z}{covariates for testing (e.g., genotype)}

\item{w.sqrt}{square root of weight vector}
}
\description{
SRAT with adjustment for additional covariates (testing)
}
